+++ R/cfbd_ratings.R
@@ -129,7 +129,7 @@ cfbd_rankings <- function(year, week = NULL, season_type = "regular") {
       polls <- res %>%
         httr::content(as = "text", encoding = "UTF-8") %>%
         jsonlite::fromJSON(flatten = TRUE) %>%
-        furrr::future_map_if(is.data.frame, list) %>%
+        purrr::map_if(is.data.frame, list) %>%
         dplyr::as_tibble() %>%
         tidyr::unnest(.data$polls) %>%
         tidyr::unnest(.data$ranks) %>%